refactor: folderize auth pages
