+++ src/__test__/pages/ResetPassword/index.test.jsx
@@ -5,7 +5,7 @@ import {
   waitFor,
   fireEvent,
 } from "@testing-library/react";
-import ResetPassword from "../../../pages/Auth/resetPassword";
+import ResetPassword from "../../../pages/Auth/ResetPassword";
 import React from "react";
 import { BrowserRouter } from "react-router-dom";
 import axios from "axios";
