+++ src/pages/Auth/index.js
@@ -1,3 +1,3 @@
-export { default as Login } from './login';
-export { default as Register } from './register';
-export { default as ResetPassword } from './resetPassword';
+export { default as Login } from './Login';
+export { default as Register } from './Register';
+export { default as ResetPassword } from './ResetPassword';
